Refactor Cart.vue to remove items from the cart and update total price
Improve the functionality of the Cart.vue component by refactoring the code to remove items from the cart. The code now correctly removes the selected furniture items from the cart and updates the total price to 0. This change ensures that the cart reflects the correct items and prices after removal.

If Item checkout, all data is inserted in order table then clear the cart table.
If item deleted by manually, it deleted from cart individually