Refactor CatalogController to update cart functionality
This commit refactors the CatalogController.php file to improve the cart functionality. It updates the logic for adding items to the cart, ensuring that the correct furniture is selected based on the code and color. refactor with  different approach, instead of passing uuid from catalog, it passes code and color , and  check the UUID in controller